+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/model/KafkaBrokerConfigurationBuilderTest.java
@@ -36,22 +36,13 @@ import io.strimzi.operator.common.Reconciliation;
 import io.strimzi.operator.common.model.cruisecontrol.CruiseControlConfigurationParameters;
 import io.strimzi.test.annotations.ParallelSuite;
 import io.strimzi.test.annotations.ParallelTest;
-import org.hamcrest.Description;
-import org.hamcrest.Matcher;
-import org.hamcrest.TypeSafeMatcher;
 
-import java.io.PrintWriter;
-import java.io.StringWriter;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.TreeSet;
 
-import static io.strimzi.operator.cluster.model.KafkaBrokerConfigurationBuilderTest.IsEquivalent.isEquivalent;
+import static io.strimzi.operator.cluster.TestUtils.IsEquivalent.isEquivalent;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
@@ -2722,76 +2713,5 @@ public class KafkaBrokerConfigurationBuilderTest {
         assertThat(configuration, isEquivalent("node.id=2",
                 "metadata.log.dir=/my/kraft/metadata/kafka-log2"
         ));
-    }
-
-    static class IsEquivalent extends TypeSafeMatcher<String> {
-        private final List<String> expectedLines;
-
-        public IsEquivalent(String expectedConfig) {
-            super();
-            this.expectedLines = ModelUtils.getLinesWithoutCommentsAndEmptyLines(expectedConfig);
-        }
-
-        public IsEquivalent(List<String> expectedLines) {
-            super();
-            this.expectedLines = expectedLines;
-        }
-
-        @Override
-        protected boolean matchesSafely(String config) {
-            List<String> actualLines = ModelUtils.getLinesWithoutCommentsAndEmptyLines(config);
-
-            return expectedLines.containsAll(actualLines) && actualLines.containsAll(expectedLines);
-        }
-
-        private String getLinesAsString(Collection<String> configLines)   {
-            StringWriter stringWriter = new StringWriter();
-            PrintWriter writer = new PrintWriter(stringWriter);
-
-            for (String line : configLines) {
-                writer.println(line);
-            }
-
-            return stringWriter.toString();
-        }
-
-        @Override
-        public void describeTo(Description description) {
-            description.appendText(getLinesAsString(new TreeSet<>(expectedLines)));
-        }
-
-        @Override
-        protected void describeMismatchSafely(String item, Description mismatchDescription) {
-            printDiff(item, mismatchDescription);
-        }
-
-        private void printDiff(String item, Description mismatchDescription)    {
-            List<String> actualLines = ModelUtils.getLinesWithoutCommentsAndEmptyLines(item);
-            List<String> actualLinesDiff = new ArrayList<>(actualLines);
-            actualLinesDiff.removeAll(expectedLines);
-            List<String> expectedLinesDiff = new ArrayList<>(expectedLines);
-            expectedLinesDiff.removeAll(actualLines);
-
-            mismatchDescription
-                    .appendText(" was: \n")
-                    .appendText(getLinesAsString(new TreeSet<>(ModelUtils.getLinesWithoutCommentsAndEmptyLines(item))))
-                    .appendText("\n\n")
-                    .appendText(" wrong lines in expected:\n")
-                    .appendText(getLinesAsString(expectedLinesDiff))
-                    .appendText("\n\n")
-                    .appendText(" Wrong lines in actual:\n")
-                    .appendText(getLinesAsString(actualLinesDiff))
-                    .appendText("\n\nOriginal value: \n")
-                    .appendText(item)
-                    .appendText("\n\n");
-        }
-
-        public static Matcher<String> isEquivalent(String expectedConfig) {
-            return new IsEquivalent(expectedConfig);
-        }
-
-        public static Matcher<String> isEquivalent(String... expectedLines) {
-            return new IsEquivalent(asList(expectedLines));
-        }
     }
 }